my_agent/deploy_agent.py
#!/usr/bin/env python3
"""
Deploy the Vibe Travel Agent to Google Cloud Vertex AI

This script deploys agent2.py to Vertex AI, enabling:
- Real-time Google Search
- All sub-agents (location finder, flights, hotels, etc.)
- Google Sheets/Docs export functionality
- Production-ready deployment

Prerequisites:
- Google Cloud project created
- Vertex AI API enabled
- gcloud CLI authenticated
- GOOGLE_CLOUD_PROJECT and GOOGLE_CLOUD_LOCATION in .env
"""

import os
import sys
from dotenv import load_dotenv

# Load environment variables
load_dotenv(dotenv_path='../.env')

def check_prerequisites():
    """Check if all prerequisites are met."""
    print("="*60)
    print("🔍 Checking Prerequisites...")
    print("="*60)
    
    errors = []
    warnings = []
    
    # Check environment variables
    PROJECT_ID = os.getenv("GOOGLE_CLOUD_PROJECT")
    LOCATION = os.getenv("GOOGLE_CLOUD_LOCATION")
    MODEL_ID = os.getenv("MODEL_ID")
    
    if not PROJECT_ID:
        errors.append("❌ GOOGLE_CLOUD_PROJECT not set in .env file")
    else:
        print(f"✅ Project ID: {PROJECT_ID}")
    
    if not LOCATION:
        errors.append("❌ GOOGLE_CLOUD_LOCATION not set in .env file")
    else:
        print(f"✅ Location: {LOCATION}")
    
    if not MODEL_ID:
        warnings.append("⚠️  MODEL_ID not set, will use default")
    else:
        print(f"✅ Model ID: {MODEL_ID}")
    
    # Check gcloud CLI
    import subprocess
    try:
        result = subprocess.run(
            ['gcloud', 'config', 'get-value', 'project'],
            capture_output=True,
            text=True,
            timeout=10
        )
        if result.returncode == 0 and result.stdout.strip():
            gcloud_project = result.stdout.strip()
            print(f"✅ gcloud CLI configured (project: {gcloud_project})")
            if PROJECT_ID and gcloud_project != PROJECT_ID:
                warnings.append(f"⚠️  gcloud project ({gcloud_project}) != .env project ({PROJECT_ID})")
        else:
            errors.append("❌ gcloud CLI not configured. Run: gcloud auth login")
    except FileNotFoundError:
        errors.append("❌ gcloud CLI not installed. Install from: https://cloud.google.com/sdk/docs/install")
    except Exception as e:
        errors.append(f"❌ Error checking gcloud: {e}")
    
    # Check if agent2.py exists
    if not os.path.exists('agent2.py'):
        errors.append("❌ agent2.py not found. Run this script from the my_agent directory.")
    else:
        print("✅ agent2.py found")
    
    # Check required packages
    try:
        import google.cloud.aiplatform
        print("✅ google-cloud-aiplatform installed")
    except ImportError:
        errors.append("❌ google-cloud-aiplatform not installed. Run: pip install google-cloud-aiplatform")
    
    try:
        import google.adk
        print("✅ google-adk installed")
    except ImportError:
        errors.append("❌ google-adk not installed. Run: pip install google-adk")
    
    # Print results
    print()
    if warnings:
        print("⚠️  Warnings:")
        for warning in warnings:
            print(f"   {warning}")
        print()
    
    if errors:
        print("❌ Errors found:")
        for error in errors:
            print(f"   {error}")
        print()
        print("Please fix the errors above before deploying.")
        return False
    
    print("✅ All prerequisites met!")
    return True


def deploy_agent():
    """Deploy the agent to Vertex AI."""
    print()
    print("="*60)
    print("🚀 Deploying Agent to Vertex AI...")
    print("="*60)
    print()
    
    try:
        from google.cloud import aiplatform
        from agent2 import root_agent
        
        PROJECT_ID = os.getenv("GOOGLE_CLOUD_PROJECT")
        LOCATION = os.getenv("GOOGLE_CLOUD_LOCATION", "us-central1")
        AGENT_NAME = os.getenv("AGENT_NAME", "vibe-travel-agent")
        
        # Initialize Vertex AI
        print(f"📤 Initializing Vertex AI...")
        print(f"   Project: {PROJECT_ID}")
        print(f"   Location: {LOCATION}")
        print(f"   Agent: {AGENT_NAME}")
        
        aiplatform.init(
            project=PROJECT_ID,
            location=LOCATION
        )
        
        print("\n✅ Vertex AI initialized")
        
        # Deploy the agent
        print(f"\n📦 Deploying agent '{AGENT_NAME}'...")
        print("   This may take 2-5 minutes...")
        
        # Note: The actual deployment method depends on Google ADK version
        # This is a placeholder - check Google ADK docs for latest deployment API
        print()
        print("⚠️  DEPLOYMENT METHOD:")
        print()
        print("Google ADK deployment is typically done through the Vertex AI Console")
        print("or using the Vertex AI SDK. Here are the steps:")
        print()
        print("METHOD 1: Using Vertex AI Console (Recommended)")
        print("1. Go to: https://console.cloud.google.com/vertex-ai/agents")
        print("2. Click 'Create Agent'")
        print("3. Upload your agent2.py file")
        print("4. Configure the agent settings")
        print("5. Deploy")
        print()
        print("METHOD 2: Using gcloud CLI")
        print("Run the following command:")
        print()
        print(f"gcloud ai agents deploy \\")
        print(f"    --agent-file=agent2.py \\")
        print(f"    --display-name={AGENT_NAME} \\")
        print(f"    --project={PROJECT_ID} \\")
        print(f"    --location={LOCATION}")
        print()
        print("METHOD 3: Programmatic Deployment")
        print("Check the latest Google ADK documentation for the deployment API:")
        print("https://cloud.google.com/vertex-ai/docs/generative-ai/adk/deploy")
        print()
        
        # For now, we'll provide instructions
        print("="*60)
        print("📋 NEXT STEPS:")
        print("="*60)
        print()
        print("1. Your agent is ready for deployment in agent2.py")
        print("2. Choose a deployment method above")
        print("3. After deployment, test with: python test_deployed_agent.py")
        print()
        print("Agent Configuration:")
        print(f"   - Name: {root_agent.name}")
        print(f"   - Model: {root_agent.model}")
        print(f"   - Tools: {len(root_agent.tools)} tools configured")
        print(f"   - Google Search: ✅ Enabled")
        print()
        
        return True
        
    except ImportError as e:
        print(f"❌ Import error: {e}")
        print("   Make sure all required packages are installed:")
        print("   pip install google-cloud-aiplatform google-adk")
        return False
    except Exception as e:
        print(f"❌ Deployment error: {e}")
        import traceback
        traceback.print_exc()
        return False


def create_test_script():
    """Create a test script for the deployed agent."""
    print()
    print("="*60)
    print("📝 Creating Test Script...")
    print("="*60)
    
    test_script = '''#!/usr/bin/env python3
"""
Test the deployed Vibe Travel Agent on Vertex AI

This script tests the deployed agent with real-time Google Search capabilities.
"""

import os
from dotenv import load_dotenv
from google.cloud import aiplatform

# Load environment variables
load_dotenv(dotenv_path='../.env')

def test_deployed_agent():
    """Test the deployed agent."""
    PROJECT_ID = os.getenv("GOOGLE_CLOUD_PROJECT")
    LOCATION = os.getenv("GOOGLE_CLOUD_LOCATION", "us-central1")
    AGENT_NAME = os.getenv("AGENT_NAME", "vibe-travel-agent")
    
    print("="*60)
    print("🧪 Testing Deployed Agent")
    print("="*60)
    print(f"\\nProject: {PROJECT_ID}")
    print(f"Location: {LOCATION}")
    print(f"Agent: {AGENT_NAME}\\n")
    
    # Initialize Vertex AI
    aiplatform.init(project=PROJECT_ID, location=LOCATION)
    
    # TODO: Load and test the deployed agent
    # The exact method depends on Google ADK version
    print("⚠️  Please check Google ADK documentation for the latest API")
    print("   to load and invoke deployed agents.")
    print()
    print("Example test query:")
    print("'Find me a 5-day beach vacation from New York with $3000 budget'")
    
if __name__ == "__main__":
    test_deployed_agent()
'''
    
    try:
        with open('test_deployed_agent.py', 'w') as f:
            f.write(test_script)
        print("✅ Created test_deployed_agent.py")
        return True
    except Exception as e:
        print(f"❌ Error creating test script: {e}")
        return False


def main():
    """Main function."""
    print()
    print("╔" + "="*58 + "╗")
    print("║" + " "*10 + "Vibe Travel Agent - Vertex AI Deployment" + " "*8 + "║")
    print("╚" + "="*58 + "╝")
    print()
    
    # Check prerequisites
    if not check_prerequisites():
        print()
        print("💡 To fix setup issues:")
        print("   1. Follow VERTEX_AI_DEPLOYMENT_GUIDE.md")
        print("   2. Ensure Google Cloud project is created")
        print("   3. Run: gcloud auth login")
        print("   4. Update .env with GOOGLE_CLOUD_PROJECT and GOOGLE_CLOUD_LOCATION")
        sys.exit(1)
    
    # Ask for confirmation
    print()
    print("⚠️  You are about to deploy to Google Cloud Vertex AI.")
    print("   This may incur costs (though free tier is usually sufficient for testing).")
    print()
    
    if os.isatty(sys.stdin.fileno()):
        response = input("Continue with deployment? (yes/no): ").strip().lower()
        if response not in ['yes', 'y']:
            print("❌ Deployment cancelled.")
            sys.exit(0)
    else:
        print("ℹ️  Running in non-interactive mode. Use the manual deployment methods.")
    
    # Deploy
    success = deploy_agent()
    
    if success:
        # Create test script
        create_test_script()
        
        print()
        print("="*60)
        print("✅ Deployment Preparation Complete!")
        print("="*60)
        print()
        print("📚 Documentation:")
        print("   - See: VERTEX_AI_DEPLOYMENT_GUIDE.md")
        print("   - Google ADK Docs: https://cloud.google.com/vertex-ai/docs/generative-ai/adk")
        print()
        print("🎯 Your agent is ready to deploy!")
        print()
    else:
        print()
        print("❌ Deployment preparation failed.")
        print("   Check errors above and try again.")
        sys.exit(1)


if __name__ == "__main__":
    main()

