requirements.txt
# Google ADK and AI Platform
google-adk>=1.18.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.125.0

# Google API Client
google-api-python-client>=2.157.0
google-auth>=2.0.0
google-auth-httplib2>=0.2.0

# Environment and Configuration
python-dotenv>=1.0.0

# Google Cloud Services (for tools.py - Google Sheets and Docs)
# Note: Service account credentials are part of google-auth (already included above)
google-cloud-storage>=3.0.0

# Optional: For test2.py if using Tavily for web search
# Uncomment if you want to use Tavily instead of google_search
# tavily-python>=0.3.0

# Standard dependencies
requests>=2.32.0
pydantic>=2.0.0

# Additional dependencies that might be needed
protobuf>=3.20.0
grpcio>=1.33.0

